requirements.txt
python-dotenv

requests
python-dotenv
beautifulsoup4
openai
certifi
PyPDF2
streamlit